apps/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6

sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.14.0

bcrypt==4.3.0
authlib==1.3.1
email-validator==2.2.0
pydantic==2.8.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.20

loguru==0.7.2
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.28.1

black==24.8.0
isort==5.13.2

requests==2.32.4
google-genai==1.31.0